HBASE-23330: Add an un-authenticated cluster ID fetch endpoint
Adds a new endpoint (/clusterId.jsp) that returns the cluster ID,
if one exists. Clients that relied on ZK earlier now can fetch
that data from this http end point. Since connection establishment
requires a token, the end point must not be authenticated (auth
happens as a part of connection handshake). For this, the patch
adds the ability to bypass the authentication filter for specific
pre determined end points.